% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs.R
\name{cbs}
\alias{cbs}
\title{CBS scrape}
\usage{
cbs(id)
}
\arguments{
\item{id}{an integer}
}
\value{
if given a valid id, a data frame with player name, dob, id, etc.
}
\description{
gets the name, dob, etc for a player, given
a cbsid value (an integer, generally between 1263000 and 2218000)
}

