% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs.R
\name{cbs_bulk}
\alias{cbs_bulk}
\title{Bulk CBS player download}
\usage{
cbs_bulk()
}
\value{
a data frame with all of the MLB players returned by the `/fantasy/player`
CBS API endpoint
}
\description{
Bulk CBS player download
}

