% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs.R
\name{cbs_scrape}
\alias{cbs_scrape}
\title{Scrape new ids for CBS}
\usage{
cbs_scrape()
}
\value{
writes data to data-raw/cbsids.csv. returns 'OK' if function completes
}
\description{
hits the cbs player endpoint, then calls each playerid
to bring down additional detail (DOB, etc) from the cbs player pagers
}

