% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{list2mat}
\alias{list2mat}
\title{list to matrix}
\usage{
list2mat(inList)
}
\arguments{
\item{inList}{list}
}
\value{
matrix
}
\description{
list to matrix
}
\references{
http://stackoverflow.com/a/24609337/561698
}

