% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb.R
\name{mlb}
\alias{mlb}
\title{MLB}
\usage{
mlb(id)
}
\arguments{
\item{id}{an integer}
}
\value{
if given a valid id, a data frame with player name, dob, id, etc.
}
\description{
gets the name and other demographic data for a player, given
a mlbid value (an integer between 110001 and something roughly in the ~ 700000s)
}
\examples{
ex <- mlb(110001)
}

