% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_new_ids}
\alias{scrape_new_ids}
\title{Scrape new ids}
\usage{
scrape_new_ids(id_type = "mlb", exclude_known = TRUE, verbose = TRUE,
  expand_range = 200, backtrack_range = 0)
}
\arguments{
\item{id_type}{c('mlb', 'yahoo')}

\item{exclude_known}{should we ignore the players who have already have
records?  default is TRUE.}

\item{verbose}{output current player name to the console.  default is TRUE}

\item{expand_range}{at the close, we'll expand the number of ids to check
by some fixed number, to make sure that we are looking high enough
to get new ones.  default is 200.}
}
\value{
'OK' if function completes
}
\description{
a utility function that looks for newly added players from
one of the fantasy providers
}

