% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo.R
\name{yahoo}
\alias{yahoo}
\title{yahoo}
\usage{
yahoo(id)
}
\arguments{
\item{id}{}
}
\value{
a data frame
}
\description{
gets the name and other demographic data for a player, given
a mlbid value (an integer between 1 and roughly 10K)
}

